% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{beta_hat}
\alias{beta_hat}
\title{Beta Hat}
\usage{
beta_hat()
}
\value{
A \code{matrix} containing the resulting beta hat value(s).
#' @author Rukesh Gusain, Michael Zirpoli, Erica Maul
}
\description{
Calculate the optimized beta hat value.
}
