% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(data = data(y, X), alpha, B = 20)
}
\arguments{
\item{alpha}{A \code{numeric} inputted by the user to find desired quantiles.}

\item{B}{A \code{numeric} describing the desired amount of bootstrap iterations.}

\item{X}{The initial \code{matrix} to be used in the function.}

\item{y}{A \code{vector} of y values.}
}
\value{
A \code{matrix} containing the calculated quantiles.
#' @author Rukesh Gusain, Michael Zirpoli, Erica Maul
}
\description{
Use a bootstrap approach to optimize the logistic regression and begin to find confidence intervals.
}
