% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSingleTests.R
\name{CalcSingleTests}
\alias{CalcSingleTests}
\title{Print Single Results}
\usage{
CalcSingleTests(model, xtrain, xtest)
}
\arguments{
\item{model, }{xtrain, xtest}
}
\description{
Prints the test results for single variables
xtrain and xtest are dataframes
}
\examples{
CalcSingleTests()
}
\keyword{tests}
\keyword{wald}

