% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Char2Factors.R
\name{Char2Factors}
\alias{Char2Factors}
\title{Convert character columns into factors}
\usage{
Char2Factors(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
This function converts character columns to factors
x is a data frame
}
\examples{
Char2Factors()
}
\keyword{factors}

