% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls.R
\name{cls}
\alias{cls}
\title{Clear Console}
\usage{
cls()
}
\arguments{
\item{x}{Clear automatically}
}
\description{
Clears console for 100 lines to make output easier to read
}
\examples{
cls()
}
\keyword{clear}
\keyword{console}

