% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{data}
\alias{data}
\title{Data}
\usage{
data(y, X)
}
\arguments{
\item{y}{A \code{vector} of y values to be used in the function.}

\item{X}{A \code{matrix} to be used in the function.}
}
\value{
A \code{dataframe} of the inputted data.
#' @author Rukesh Gusain, Michael Zirpoli, Erica Maul
}
\description{
Create a data frame from the X matrix and y vector.
}
