% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoPlots.R
\name{DoPlots}
\alias{DoPlots}
\title{Plot information about predictions on test set}
\usage{
DoPlots(predictions, performance1, ytest)
}
\arguments{
\item{predictions, }{performance1, ytest}
}
\description{
Plots prediction information
predictions are from a model predict object
performance1 is the output from the LossInfo function
ytest is a factor with two levels
}
\examples{
DoPlots()
}
\keyword{plots}

