% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorID.R
\name{FactorID}
\alias{FactorID}
\title{Detect Factors}
\usage{
FactorID(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
This function determines which columns in a data frame x are nominal
}
\examples{
FactorID()
}
\keyword{factors}

