% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findK.R
\name{findK}
\alias{findK}
\title{Find K Segments}
\usage{
findK(X, Y, col, maxK, colName)
}
\arguments{
\item{X, }{Y, col, maxK, colName}
}
\description{
Segments numerical variables into categories using k-means
X is a data frame, Y is a response that is a factor with two levels
}
\examples{
findK()
}
\keyword{K-means}
\keyword{segmentation}

