% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClustStats.R
\name{GetClustStats}
\alias{GetClustStats}
\title{Get the cluster statistics}
\usage{
GetClustStats(Clusters, x, k)
}
\arguments{
\item{Clusters, }{x, k}
}
\description{
Table of within cluster statistics
Cluster is a vector of integers indicating clusters
x is a vector and k is a non-negative integer
}
\examples{
GetClustStats()
}
\keyword{K-means}
\keyword{plots}
\keyword{segmentation}

