% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInterNomBarPlots.R
\name{GetInterNomBarPlots}
\alias{GetInterNomBarPlots}
\title{Interleaved Barplots for a dataframe (> 1 column)}
\usage{
GetInterNomBarPlots(data, Response)
}
\arguments{
\item{response, }{x}
}
\description{
For a multiple class response ( > 1) where x is a data frame and 
response is a factor vector with two levels
}
\examples{
GetInterNomBarPlots()
}
\keyword{barplots}
\keyword{factor}
\keyword{interleaved}
\keyword{ordered}

