% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNomBarPlots.R
\name{GetNomBarPlots}
\alias{GetNomBarPlots}
\title{Barplots for a dataframe (> 1 column)}
\usage{
GetNomBarPlots(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
Returns a barplot, x is a data frame
}
\examples{
GetNomBarPlots()
}
\keyword{barplots}
\keyword{factor}
\keyword{ordered}

