% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient Descent Algorithm}
\usage{
gradient_descent(x, y, theta, lr, max_iter, threshold, regu_factor,
  regu_method, epsilon)
}
\arguments{
\item{x}{Matrix with the predictor variables}

\item{y}{Matrix with the target variable}

\item{theta}{Matrix with the theta's values for sigmoid function}

\item{lr}{Learning Rate for Gradient Descent Algorithm}

\item{max_iter}{Number of iterations for Gradient Descent Algorithm}

\item{threshold}{Threshold for Gradient Descent Algorithm}

\item{regu_factor}{Regularization Factor for Ridge Regularization Regression}

\item{regu_method}{Regularization Method. At this moment, only "ridge" available}

\item{epsilon}{Numeric value}
}
\value{
Returns the optimal theta's values and the cost function values
}
\description{
Gradient Descent Algorithm to get the best values of theta for logistic regression
}
\author{
Alberto Almuiña
}
