% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{initial_beta}
\alias{initial_beta}
\title{Initial Beta}
\usage{
initial_beta(X, y)
}
\arguments{
\item{X}{An initial \code{matrix} to be used in the function.}

\item{y}{A \code{vector} of y values to be used in the function. simulations.}
}
\value{
A \code{matrix} containing the initial beta value(s).
#' @author Rukesh Gusain, Michael Zirpoli, Erica Maul
}
\description{
Calculate an initial beta value that will be used as the parameter of the optimization function for beta hat, using a least squares equation.
}
