% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterleavedRelever.R
\name{InterleavedRelever}
\alias{InterleavedRelever}
\title{Interleaved Relever}
\usage{
InterleavedRelever(x, Response, colName)
}
\arguments{
\item{x, }{Response, colName}
}
\description{
Interleaved, relevels factors so they can be plotted from highest to lowest frequency
x is a data frame
Response is a factor with two levels
colName is the column to be compared to the response
}
\examples{
InterleavedRelever()
}
\keyword{factors}
\keyword{interleaved}
\keyword{plots}
\keyword{pretty}

