% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{logistic_curve}
\alias{logistic_curve}
\title{Logistic Curve}
\usage{
logistic_curve(data = data(y, X), predictor, beta_hat = beta_hat())
}
\arguments{
\item{data}{The \code{dataframe} of the data to use for the plot.}

\item{predictor}{A \code{character} to assign the column name corresponding to the predictor.}

\item{beta_hat}{A \code{matrix} of the optimized beta values.}
}
\value{
A plot depicting a fitted logistic curve to the actual values.
#' @author Rukesh Gusain, Michael Zirpoli, Erica Maul
}
\description{
Plot the logistic regression of the data.
}
