% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg_comprobations.R
\name{logistic_reg_comprobations}
\alias{logistic_reg_comprobations}
\title{Logistic Regression Comprobations}
\usage{
logistic_reg_comprobations(target, lr, max_iter, regu_factor, threshold,
  probs, epsilon)
}
\arguments{
\item{target}{Variable target for logistic regression}

\item{lr}{Learning rate for gradient descent algorithm}

\item{max_iter}{Number of iterations for gradient descent algorithm}

\item{regu_factor}{Regularization factor for Ridge Regularization}

\item{threshold}{Threshold to stop gradient descent algorithm base on the value of the cost function}

\item{probs}{If the results must be given like probabilities or not.}

\item{epsilon}{Numeric parameter for computation}
}
\value{
Returns the mean values for the metric selected
}
\description{
Comprobations for the "logistic_regression" function parameters.
}
\author{
Alberto Almuiña
}
