% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic Regression}
\usage{
logistic_regression(train, test, target, lr = 0.001, max_iter = 1e+05,
  probs = F, threshold = NULL, regu_factor = NULL,
  regu_method = "ridge", epsilon = 1e-05, ...)
}
\arguments{
\item{train}{Matrix with the predictor variables}

\item{test}{Matrix with the target variable}

\item{target}{Numeric value of the target variable}

\item{lr}{Learning Rate for Gradient Descent Algorithm}

\item{max_iter}{Number of iterations for Gradient Descent Algorithm}

\item{probs}{If the result must be given in probabilities or not.}

\item{threshold}{Threshold for gradient descent algorithm}

\item{regu_factor}{Regularization Factor for Ridge Regularization Regression}

\item{regu_method}{Regularization Method. At this moment, only "ridge" available}

\item{epsilon}{Numeric value}

\item{...}{Other possible arguments}
}
\value{
Returns a list with the predicted data, confussion matrix, confussion matrix plot and optimal theta values
}
\description{
Logistic Regression for Binary Classification Problems.
}
\author{
Alberto Almuiña
}
