% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticGLM.R
\name{LogisticGLM}
\alias{LogisticGLM}
\title{Logistic Regression}
\usage{
LogisticGLM(xtrain, ytrain, xtest, ytest)
}
\arguments{
\item{xtrain, }{ytrain, xtest, ytest}
}
\description{
This function calculates and plots Logistic Regression via glm()
xtrain and xtest are data frames,
}
\examples{
LogisticGLM()
}
\keyword{Logistic}
\keyword{Regression}

