% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLossBinary.R
\name{LogLossBinary}
\alias{LogLossBinary}
\title{Log Loss Binary}
\usage{
LogLossBinary(y, yhat, eps = 1e-15)
}
\arguments{
\item{y, }{yhat}
}
\description{
Calculates Log Loss Binary
y is the known response, yhat is the predicted
}
\examples{
LogLossBinary()
}
\keyword{loss}

