% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{loss_func}
\alias{loss_func}
\title{Loss Function}
\usage{
loss_func(y, X, beta = initial_beta(X, y))
}
\arguments{
\item{y}{A \code{vector} of y values to be used in the function.}

\item{X}{An initial \code{matrix} to be used in the function.}

\item{beta}{The \code{matrix} of initial beta values produced in the Initial Beta function.}
}
\value{
A \code{numeric} of a value to be optimized.
#' @author Rukesh Gusain, Michael Zirpoli, Erica Maul
}
\description{
First, solve for p as a step in finding the beta hat optimization. Second, calculate all of the beta hat function before optimization.
}
