% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossInfo.R
\name{LossInfo}
\alias{LossInfo}
\title{Calculate and print loss}
\usage{
LossInfo(model, predictions, xtest, ytest)
}
\arguments{
\item{model, }{predictions, xtest, ytest}
}
\description{
Prints loss function information
}
\examples{
LossInfo()
}
\keyword{loss}

