% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{metrics}
\alias{metrics}
\title{Metrics}
\usage{
metrics(X = X, beta = initial_beta(X, y), beta_hat = beta_hat(), cutoff = 0.5)
}
\arguments{
\item{X}{A \code{matrix} from which data will be pulled to fit the model.}

\item{beta}{The \code{matrix} of initial beta values produced in the Initial Beta function.}

\item{beta_hat}{A \code{matrix} of the optimized beta values.}

\item{cutoff}{A \code{numeric} specified by the user for a cutoff value.}
}
\value{
A \code{vector} of the calculated metrics.
#' @author Rukesh Gusain, Michael Zirpoli, Erica Maul
}
\description{
Calculate a number of statistical metrics from a confusion matrix comparing predicted and actual data.
}
