% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist, }{xtest, ytest}
}
\description{
Plots multiple ggplots on on page
Input is a list of ggplots
}
\examples{
multiplot()
}
\keyword{model}

