% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{newton_method.logistic_regression}
\alias{newton_method.logistic_regression}
\title{Performs Newton Method on a `"logistic_regression"` object to find optimal parameters.}
\usage{
\method{newton_method}{logistic_regression}(lr)
}
\arguments{
\item{lr}{Instance of class \code{\link{logistic_regression}}.}
}
\value{
Column vector of coefficients optimized via Newton's method.
}
\description{
Performs Newton Method on a `"logistic_regression"` object to find optimal parameters.
}
