% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numLvls.R
\name{numLvls}
\alias{numLvls}
\title{Count number of factor levels}
\usage{
numLvls(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
Calculates the number of levels in a categorical variable and subtracts 1
x is a data frame
}
\examples{
numLvls()
}
\keyword{factors}

