% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotClusters.R
\name{PlotClusters}
\alias{PlotClusters}
\title{Plot Clusters}
\usage{
PlotClusters(x1, Cluster, colName)
}
\arguments{
\item{x1, }{Cluster, colName}
}
\description{
Plots the k-means++ clustering in two ways
x1 is a vector, Cluster is the cluster output from kmeans++
colName is the name of the column that has been clustered
}
\examples{
PlotClusters()
}
\keyword{K-means}
\keyword{plots}
\keyword{segmentation}

