% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{predict.logistic_regression}
\alias{predict.logistic_regression}
\title{Predicts labels of a test set using a 0.5 cutoff
If \eqn{\sigma({\bf{x}}_i^{\text{test}} \beta) > 0.5} then observation \eqn{i} is assigned
class 1. Otherwise it is assigned class 0.}
\usage{
\method{predict}{logistic_regression}(lr, xtest)
}
\arguments{
\item{lr}{Instance of class \code{\link{logistic_regression}}.}

\item{xtest}{Matrix of dimension (number of test observations, number of features + 1) containing a
testing observation in every row and a feature in every column. Notice that the first column must contain
only 1s to fit the intercept.}
}
\value{
Column vector of dimension (number of test observations, 1) with a label of `0` or `1` for every
test observation.
}
\description{
Predicts labels of a test set using a 0.5 cutoff
If \eqn{\sigma({\bf{x}}_i^{\text{test}} \beta) > 0.5} then observation \eqn{i} is assigned
class 1. Otherwise it is assigned class 0.
}
