% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict}
\usage{
predict(x_train, x_test, theta)
}
\arguments{
\item{x_train}{Matrix with the predictor variables of the training dataset}

\item{x_test}{Matrix with the predictor variables of the test dataset}

\item{theta}{Matrix with the theta's values for sigmoid function}
}
\value{
Returns the predicted values for logistic regression
}
\description{
Predict for Logistic Regression
}
\author{
Alberto Almuiña
}
