% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{print.logistic_regression}
\alias{print.logistic_regression}
\title{Pretty-Print method for S3 Object `"logistic_regression"`.}
\usage{
\method{print}{logistic_regression}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Pretty-Print method for S3 Object `"logistic_regression"`.
}
