% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintModelInfo.R
\name{PrintModelInfo}
\alias{PrintModelInfo}
\title{Print Model Info}
\usage{
PrintModelInfo(model, xtest, ytest)
}
\arguments{
\item{model, }{xtest, ytest}
}
\description{
Prints model information
model is a model object, xtest is a data frame, ytest is a factor with two levels
}
\examples{
PrintModelInfo()
}
\keyword{model}

