% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readkeygraph.R
\name{readkeygraph}
\alias{readkeygraph}
\title{Readkey Graph}
\usage{
readkeygraph(prompt)
}
\arguments{
\item{prompt}{Presstheanykey}
}
\description{
For graph key press interactivity
}
\examples{
readkeygraph()
}
\keyword{presstheanykey}

