% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relever.R
\name{Relever}
\alias{Relever}
\title{Relever}
\usage{
Relever(x, colName)
}
\arguments{
\item{x, }{colName}
}
\description{
Relevels factors so they can be plotted from highest to lowest frequency
x is a factor, colName is its name
}
\examples{
Relever()
}
\keyword{factors}
\keyword{plots}
\keyword{pretty}

