% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnTables.R
\name{ReturnTables}
\alias{ReturnTables}
\title{Print Segmentation Diagnostics}
\usage{
ReturnTables(prb, ytest, ZO1, LL1, AU1, Ranges)
}
\arguments{
\item{prb, }{ytest, ZO1, LL1, AU1, Ranges}
}
\description{
Tables for segmentation diagnostics
prb is a probability
ytest is a factor with two levels
Z01, LL1, AU1 are outputs from LossInfo
}
\examples{
ReturnTables()
}
\keyword{K-means}
\keyword{plots}
\keyword{segmentation}

