% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetClusterLevels.R
\name{SetClusterLevels}
\alias{SetClusterLevels}
\title{User Set Cluster Levels}
\usage{
SetClusterLevels(X, ClusterLevels, col)
}
\arguments{
\item{X, }{ClusterLevels, col}
}
\description{
Allows user to choose a segmentation found with k-means
X is a data frame, ClusterLevels is a list of clusters, 
col is the column number of the column being clustered
}
\examples{
SetClusterLevels()
}
\keyword{K-means}
\keyword{segmentation}

