% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{Values to compute.}
}
\value{
Returns the values of the sigmoid function.
}
\description{
Sigmoid function.
}
\author{
Alberto Almuiña
}
