% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTestsNom.R
\name{SingleTestsNom}
\alias{SingleTestsNom}
\title{Wald tests for Single Categorical Variables}
\usage{
SingleTestsNom(x, model)
}
\arguments{
\item{x, }{model A data frame and a model}
}
\description{
Conducts Wald tests on single categorical variables
x is a data frame, model is a model object
}
\examples{
SingleTestsNom()
}
\keyword{tests}
\keyword{wald}

