% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTestsNum.R
\name{SingleTestsNum}
\alias{SingleTestsNum}
\title{Wald tests for Single Numerical Variables}
\usage{
SingleTestsNum(x, model)
}
\arguments{
\item{x, }{model A data frame and a model}
}
\description{
Conducts Wald tests on single numerical variables
}
\examples{
SingleTestsNum()
}
\keyword{tests}
\keyword{wald}

