% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitData.R
\name{SplitData}
\alias{SplitData}
\title{Split the data}
\usage{
SplitData(X, Y)
}
\arguments{
\item{X, }{Y}
}
\description{
Randomly splits the data in half
}
\examples{
SplitData()
}
\keyword{K-means}
\keyword{segmentation}

