% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarSelection.R
\name{VarSelection}
\alias{VarSelection}
\title{Runs variable selection procedures}
\usage{
VarSelection(model)
}
\arguments{
\item{model}{A model}
}
\description{
Runs 5 types of variable selection and prints to console
}
\examples{
VarSelection()
}
\keyword{selection}
\keyword{variable}

