% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroOneLossRound.R
\name{ZeroOneLossRound}
\alias{ZeroOneLossRound}
\title{Rounded 0-1 Error}
\usage{
ZeroOneLossRound(y, yhat)
}
\arguments{
\item{y, }{yhat}
}
\description{
Calculates 0-1 Error after rounding predictions
}
\examples{
ZeroOneLossRound()
}
\keyword{loss}

