% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{grad_ascent.logistic_regression}
\alias{grad_ascent.logistic_regression}
\title{Performs Gradient Ascent on a `"logistic_regression"` object.
Should ideally be called internally, but can be called externally for
debugging purposes. Does not change the field `beta`.}
\usage{
\method{grad_ascent}{logistic_regression}(lr)
}
\arguments{
\item{sigmab}{}
}
\value{

}
\description{
Performs Gradient Ascent on a `"logistic_regression"` object.
Should ideally be called internally, but can be called externally for
debugging purposes. Does not change the field `beta`.
}
