% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic Regression S3 Object}
\usage{
logistic_regression(
  X,
  y,
  cost = "MLE",
  method = "BFGS",
  sigmab = 1,
  niter = 100,
  alpha = 0.1,
  gamma = 0.001
)
}
\arguments{
\item{gamma}{}
}
\value{

}
\description{
Logistic Regression S3 Object
}
