% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{run.logistic_regression}
\alias{run.logistic_regression}
\title{Runs the optimization for logistic regression}
\usage{
\method{run}{logistic_regression}(lr)
}
\arguments{
\item{lr}{}
}
\value{

}
\description{
Runs the optimization for logistic regression
}
